+++ CHANGELOG.md
@@ -21,11 +21,13 @@ This document details all notable changes to the Empty Style Guide framework. It
 
 ### Changed
 
+- Refactor demo container controller event listeners.
 - Convert default button into a mixin.
 - Revise order of the pattern demo meta and controllers to facilitate flexbox layout.
 
 ### Added
 
+- JavaScript controllers to manage the width of the pattern demo container.
 - Styles to use flexbox layout for the pattern demo meta layout.
 - Create standard and secondary button styles.
 - Reset button in the pattern demo to reset the demo container width to the viewport size or the max, whichever is smaller.